requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.2
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
uWSGI==2.0.20
